Refactored server.js
